added wrapper for std::variant
added a gc wrapper for std::variant.
unlike the other wrappers, the lockable type is the first template type parametery (instead of the last) because parameter packs must come last unless they're deducible.

all the standard utilities for using variants and accessing their contents is left in namespace std.
specializations and overloads are injected into the std namespace for use with the wrapped types.