added wrapper for std::optional
added GC::optional as a wrapper for std::optional.

during the implementation of the wrapper, i remembered a safety issue that i'd forgotten to fix back when i was working on the project more often:
the move assignment/swap stuff that take an arg by non-const reference have a potential to deadlock at one self-swap/self-(move)-assignment (unless a recursive mutex is used, which we are doing).
this'll be fixed in an upcoming update.

additionally, i've discovered a new potential problem: if a container function takes a forwarding reference to some argument, it's possible to modify it (e.g. if it's a non-const reference esp. rvalue references).
thus we need to have compile-time cases to check not only for self referencing in these cases, but also if the anonymous type is a gc wrapper type (in which case we need to lock its mutex if it's passed by non-const reference).
what makes this a tad more annoying is this can happen in template parameter packs...

so that'll be fun to fix...