fixed wrapper mutex risks
in the previous commit i mentioned the issue of std::scoped_lock trying to lock the same lock on self-(move)-assignment and swapping of wrappers.
this is indeed a problem, and this commit fixes all of those problem spots with a simple if statement.

the last commit also mentioned the possibility for modifying a wrapper passed by reference to e.g. a parameter pack and not having locked its mutex.
to this end i've added a utility type/function pair for accessing the wrapper mutex object of any type (even non-wrapper types).
for non-wrapper types this returns a reference to a __gc_fake_lock_t, which is a no-op nothing class that has the same interface as Lockable.

however, after all that was implemented and i went to add it to the locking logic, i realized it was actually a non-issue.
the wrapper types themselves never unsafely expose the ability to modify them.
additionally, if we lock the mutex before forwarding the arguments, the destination call site would deadlock on any mutation action (assuming the mutex wasn't recursive, which it is).

thus, this is not used anywhere and (unless i think up a reason to keep it) will be removed in the coming commits.