possible rpi fix
a bug came up in the rpi testing upon program termination.
not sure what the problem is, but sometimes it works and sometimes it has an assertion (one of mine) failure or a segfault.
i'm suspicious it's a lifetime issue, so i've changed all global variables to be meyers singletons.