new program term logic
i've known the problem in the rpi testing is caused by the collector thread accessing the (static singleton) gc data after they're destroyed.
this is because it's running as a detached concurrent thread, which makes it very hard to control.

this is now fixed with new logic:
the background collector thread is no longer detached - a thread handle is held to it by the disjunction container (singleton).
this is assigned by start_timed_collect() (upon its first invocation).
upon destruction of the module container, it performs special signal and flag logic to join the background collector.

this means that, upon (normal) program termination, the background collector thread is not running after the module container is destroyed.
this should (hopefully) fix the problem.

if any problem still exists, it's likely due to oddities of the local_detour logic.
this will need to be reviewed as well (i really don't like that it exists - it's kind of a hack).