another potential fix
the last patch fixed most of the issue.
the background collector running detached was causing the most consistent issues in the linux release.
however program termination still fails occassionally (only in linux).

this patch is based on the idea that the problem lies with local_detour (most likely culprit in my eye).
this patch makes local_detour an atomic object.
it's possible that the local_detour value was not being propagated to other threads due to not being synchronized.
this would only fix (non-und) cases of local(), which is more frequently used than local_handle() itself.

tbh this is a very complicated problem which might actually be impossible to do in a cross-platform, standard-conforming way.
if this turns out to be the case, i'll just remove some of the program termination cleanup guarantees.
i'd rather not have to do that, but if this doesn't fix the problem it's looking pretty grim.