finally fixed collector crash bug
turns out the last commit was very close to being a total fix for the crashes caused by the collector thread upon termination.
apparently all that was wrong was i got the new usage guards for its construction out of order with respect to the module container (which it uses).
thus it was being constructed before the module container existed (bad) and was being destroyed after the module container (also bad).
so swapping the order of construction as imposed by the usage guards fixes the problem.

also, one of the issues was actually the local_detour nullity assertion in the local_handle() accessor.
actually, that assertion is incorrect, as program termination does not imply that all threads are dead (e.g. detached threads).
this was actually erroneously causing an assertion failure (crash) on program termination from the collector thread (which is detached).
since it's not terminated immediately, its local_handle() is actually not invalidated yet, which is further guaranteed by its on_dtor event, which is set up during collector thread initialization.

all usage guards are now rolled into 2: a static usage guard and a thread_local usage guard.
this provides all the same flexibility without adding a whole bunch of unnecessary names/types.

unfortunately, there still apears to be a crash bug that's only triggered during termination, but it's caused by something else - the collector is good now (i think/hope).
the new bug is likely caused by the