test file fix
by the nine! i've discovered one of the problems with segfaulting on linux.
apparently there was a race condition with invoking thread sleep during thread destruction.
this was caused by having one of the testing types call it in the dtor, which was being invoked by the disjoin module final collection mechanism upon program termination.
my current theory is that Windows is a lot more hand-wavy with thread lifetime on program termination and/or implements thread sleep in a (seemingly) safer way.

another issue was that some threads were being detached in the test file, and that was causing some sort of race condition something.
this was likely caused by them holding owning disjunction handles, but i didn't explore further after fixing it (it should have always been a join anyway).

these fixes don't 100% fix the segfault on termination, but it's much less frequent.
on day for sure...

fixed the gc header not including <condition_variable>.