fixed comple warnings
added some more compile warnings to the makefile and fixed all the new warning output.

still attempting to find the segfault bug.
gdb says it's a use after free error between resources allocated and deallocated by different threads.
logically this would be the background collector, but use after free would be the source thread.
this seems to imply that something in the test file is wrong, which would actually be better than the alternative.

relocated some tests to be in correct gc strategy boundaries.
there were some tests in the timed strategy segment that could actually fail due to timed collection (needed to be manual).